Remove C-style casts. See #37.
Enable -Wold-style-cast and also remove all existing C-style casts. Doing this
found multiple places where const was incorrectly being cast away, as well as
some pointless casting of size_t down to int.

+++ src/Common/Utilities/src/FileHeader.cpp
@@ -113,7 +113,7 @@ namespace BitFunnel
 
     FileHeaderTime::FileHeaderTime(std::istream& in)
     {
-        const int c_len = (int)strlen("04/20/2011 02:24:45");
+        const size_t c_len = strlen("04/20/2011 02:24:45");
         char buffer[64];
         StreamUtilities::ReadBytes(in, buffer, c_len);
         buffer[c_len] = 0;
